# apps/api/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/shopspring/decimal"
)

type Agent struct {
	ID             uuid.UUID          `json:"id"`
	Name           string             `json:"name"`
	DeveloperID    *uuid.UUID         `json:"developer_id"`
	OrganizationID *uuid.UUID         `json:"organization_id"`
	Description    *string            `json:"description"`
	Active         bool               `json:"active"`
	LastSeenAt     pgtype.Timestamptz `json:"last_seen_at"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type ApiKey struct {
	ID         uuid.UUID          `json:"id"`
	AgentID    uuid.UUID          `json:"agent_id"`
	Name       string             `json:"name"`
	Prefix     string             `json:"prefix"`
	KeyHash    string             `json:"key_hash"`
	Scopes     []string           `json:"scopes"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	LastUsedAt pgtype.Timestamptz `json:"last_used_at"`
	RevokedAt  pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Organization struct {
	ID        uuid.UUID          `json:"id"`
	Name      string             `json:"name"`
	Slug      string             `json:"slug"`
	Color     *string            `json:"color"`
	Active    bool               `json:"active"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type OrganizationInvite struct {
	ID             uuid.UUID          `json:"id"`
	OrganizationID uuid.UUID          `json:"organization_id"`
	Email          *string            `json:"email"`
	Role           string             `json:"role"`
	TokenHash      string             `json:"token_hash"`
	CreatedBy      uuid.UUID          `json:"created_by"`
	ExpiresAt      pgtype.Timestamptz `json:"expires_at"`
	AcceptedAt     pgtype.Timestamptz `json:"accepted_at"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
}

type OrganizationMember struct {
	OrganizationID uuid.UUID          `json:"organization_id"`
	UserID         uuid.UUID          `json:"user_id"`
	Role           string             `json:"role"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
}

type Project struct {
	ID             uuid.UUID          `json:"id"`
	OrganizationID uuid.UUID          `json:"organization_id"`
	Name           string             `json:"name"`
	Key            string             `json:"key"`
	Description    string             `json:"description"`
	TaskCounter    int32              `json:"task_counter"`
	Active         bool               `json:"active"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type Session struct {
	ID        uuid.UUID          `json:"id"`
	UserID    uuid.UUID          `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type Task struct {
	ID                 uuid.UUID          `json:"id"`
	Number             int32              `json:"number"`
	OrganizationID     uuid.UUID          `json:"organization_id"`
	ProjectID          uuid.UUID          `json:"project_id"`
	Title              string             `json:"title"`
	Description        string             `json:"description"`
	AcceptanceCriteria string             `json:"acceptance_criteria"`
	Status             string             `json:"status"`
	Priority           string             `json:"priority"`
	AssigneeID         *uuid.UUID         `json:"assignee_id"`
	ReporterID         *uuid.UUID         `json:"reporter_id"`
	DueDate            pgtype.Date        `json:"due_date"`
	Position           decimal.Decimal    `json:"position"`
	Blocked            bool               `json:"blocked"`
	BlockedReason      *string            `json:"blocked_reason"`
	Source             string             `json:"source"`
	ExternalRef        *string            `json:"external_ref"`
	SourceMetadata     []byte             `json:"source_metadata"`
	CompletionSummary  *string            `json:"completion_summary"`
	Version            int32              `json:"version"`
	ArchivedAt         pgtype.Timestamptz `json:"archived_at"`
	CreatedAt          pgtype.Timestamptz `json:"created_at"`
	UpdatedAt          pgtype.Timestamptz `json:"updated_at"`
}

type TaskActivity struct {
	ID        uuid.UUID          `json:"id"`
	TaskID    uuid.UUID          `json:"task_id"`
	ActorType string             `json:"actor_type"`
	ActorID   *uuid.UUID         `json:"actor_id"`
	Action    string             `json:"action"`
	OldValue  []byte             `json:"old_value"`
	NewValue  []byte             `json:"new_value"`
	Metadata  []byte             `json:"metadata"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type TaskComment struct {
	ID         uuid.UUID          `json:"id"`
	TaskID     uuid.UUID          `json:"task_id"`
	AuthorType string             `json:"author_type"`
	AuthorID   *uuid.UUID         `json:"author_id"`
	Content    string             `json:"content"`
	Metadata   []byte             `json:"metadata"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UpdatedAt  pgtype.Timestamptz `json:"updated_at"`
}

type User struct {
	ID           uuid.UUID          `json:"id"`
	Name         string             `json:"name"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"password_hash"`
	AvatarUrl    *string            `json:"avatar_url"`
	Active       bool               `json:"active"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}