# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Team-Ops API
  version: 0.1.0
  description: Self-hosted engineering board for humans and AI agents.
servers:
  - url: /v1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/User"
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/User"
...
  /auth/invite:
    get:
      tags: [Auth]
      operationId: peekInvite
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/InvitePreview"
  /organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      responses:
        "200":
          $ref: "#/components/responses/OrganizationList"
    post:
      tags: [Organizations]
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          $ref: "#/components/responses/Organization"
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      parameters:
        - in: query
          name: organizationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/ProjectList"
    post:
      tags: [Projects]
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          $ref: "#/components/responses/Project"
  /tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      parameters:
        - in: query
          name: organizationId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: projectId
          schema:
            type: string
            format: uuid
        - in: query
          name: assigneeId
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - in: query
          name: priority
          schema:
            $ref: "#/components/schemas/Priority"
        - in: query
          name: blocked
          schema:
            type: boolean
        - in: query
          name: source
          schema:
            $ref: "#/components/schemas/Source"
        - in: query
          name: updatedAfter
          schema:
            type: string
            format: date-time
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: externalRef
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/TaskList"
    post:
      tags: [Tasks]
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          $ref: "#/components/responses/Task"
        "409":
          $ref: "#/components/responses/Error"
...
  /agents:
    get:
      tags: [Agents]
      operationId: listAgents
      parameters:
        - in: query
          name: organizationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          $ref: "#/components/responses/AgentList"
    post:
      tags: [Agents]
      operationId: createAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          $ref: "#/components/responses/Agent"
...
components:
...
  schemas:
    Envelope:
      type: object
      required: [data, error, meta]
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          type: object
    ErrorBody:
      type: object
      nullable: true
      properties:
        code:
          type: string
        message:
          type: string
    TaskStatus:
      type: string
      enum: [backlog, in_progress, review, done]
    Priority:
      type: string
      enum: [low, medium, high]
    Source:
      type: string
      enum: [manual, ai, integration]
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        inviteToken: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...
    CreateOrganizationRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        color: { type: string }
...
    CreateProjectRequest:
      type: object
      required: [organizationId, name, key]
      properties:
        organizationId: { type: string, format: uuid }
        name: { type: string }
        key: { type: string }
        description: { type: string }
...
    CreateTaskRequest:
      type: object
      required: [organizationId, projectId, title]
      properties:
        organizationId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        acceptanceCriteria: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
        assigneeId: { type: string, format: uuid }
        dueDate: { type: string }
        externalRef: { type: string }
        source: { $ref: "#/components/schemas/Source" }
        sourceMetadata: { type: object }
...
    CreateAgentRequest:
      type: object
      required: [organizationId, name]
      properties:
        organizationId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    User:
      description: User
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Organization:
      description: Organization
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    OrganizationList:
      description: Organizations
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    InvitePreview:
      description: Invite preview
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Project:
      description: Project
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    ProjectList:
      description: Projects
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Task:
      description: Task
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    TaskList:
      description: Tasks
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    Agent:
      description: Agent
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    AgentList:
      description: Agents
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...